+++ crypto/tss/eddsa/frost/signer/round_2.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package eddsaSigner
+package signer
 
 import (
 	"math/big"
