+++ crypto/tss/eddsa/frost/signer/signer_test.go
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package eddsaSigner
+package signer
 
 import (
 	"math/big"
